Updated inference explanation

+++ common/file_utils.py
@@ -112,32 +112,35 @@ def get_file(fname, origin, untar=False,
     '''
 
     if download:
-        print('Downloading data from', origin)
-        global progbar
-        progbar = None
-
-        def dl_progress(count, block_size, total_size):
+        if 'modac.cancer.gov' in origin:
+            get_file_from_modac(fpath, origin)
+        else:
+            print('Downloading data from', origin)
             global progbar
-            if progbar is None:
-                progbar = Progbar(total_size)
-            else:
-                progbar.update(count * block_size)
-
-        error_msg = 'URL fetch failure on {}: {} -- {}'
-        try:
+            progbar = None
+    
+            def dl_progress(count, block_size, total_size):
+                global progbar
+                if progbar is None:
+                    progbar = Progbar(total_size)
+                else:
+                    progbar.update(count * block_size)
+    
+            error_msg = 'URL fetch failure on {}: {} -- {}'
             try:
-                urlretrieve(origin, fpath, dl_progress)
-                # fpath = wget.download(origin)
-            except URLError as e:
-                raise Exception(error_msg.format(origin, e.errno, e.reason))
-            except HTTPError as e:
-                raise Exception(error_msg.format(origin, e.code, e.msg))
-        except (Exception, KeyboardInterrupt) as e:
-            if os.path.exists(fpath):
-                os.remove(fpath)
-            raise
-        progbar = None
-        print()
+                try:
+                    urlretrieve(origin, fpath, dl_progress)
+                    # fpath = wget.download(origin)
+                except URLError as e:
+                    raise Exception(error_msg.format(origin, e.errno, e.reason))
+                except HTTPError as e:
+                    raise Exception(error_msg.format(origin, e.code, e.msg))
+            except (Exception, KeyboardInterrupt) as e:
+                if os.path.exists(fpath):
+                    os.remove(fpath)
+                raise
+            progbar = None
+            print()
 
     if untar:
         if not os.path.exists(untar_fpath):
